Creation of file + Base system
Wrote a WIP description of the module
Created the DIMENSION and DEFAULT_CELLS and P constants
Defined a function that could create a plate which can contain the cells. Each cells has it's load of informations and properties.
The cell are disposed in a hex pattern, i.e they can have up to 6 neighbours
Defined a function that returns the neighbours of a given cell. Although it's messy, because of the peculiar cases of the borders, it works fine. May need to optimise it though.